third_party/WebKit/LayoutTests/inspector/sources/debugger-pause/debugger-eval-while-paused-throws-expected.txt

Tests that evaluation in console that throws works fine when script is paused.

Set timer for test function.
Dumping console messages:

injectedFunction()
VM:4 Uncaught Error: injectedObj.func
    at Object.func (<anonymous>:4:15)
    at injectedFunction (<anonymous>:7:24)
    at eval (eval at testFunction (debugger-eval-while-…used-throws.html:18), <anonymous>:1:1)
    at testFunction (test.js:18)
func @ VM:4
injectedFunction @ VM:7
(anonymous) @ VM:1
testFunction @ test.js:18
setTimeout (async)
scheduleTestFunction @ debugger-test.js:3
(anonymous) @ VM:1
localObj.func()
test.js:15 Uncaught Error: localObj.func
    at Object.func (test.js:15)
    at eval (eval at testFunction (debugger-eval-while-…used-throws.html:18), <anonymous>:1:10)
    at testFunction (test.js:18)
func @ test.js:15
(anonymous) @ VM:1
testFunction @ test.js:18
setTimeout (async)
scheduleTestFunction @ debugger-test.js:3
(anonymous) @ VM:1
globalObj.func()
test.js:6 Uncaught Error: globalObj.func
    at Object.func (test.js:6)
    at eval (eval at testFunction (debugger-eval-while-…used-throws.html:18), <anonymous>:1:11)
    at testFunction (test.js:18)
func @ test.js:6
(anonymous) @ VM:1
testFunction @ test.js:18
setTimeout (async)
scheduleTestFunction @ debugger-test.js:3
(anonymous) @ VM:1
